DetectorForWaterTreatmentPlant のリファクタリング #41
Detector.csの形式に合わせた
